requirements.txt
discord.py>=2.0.0
aiosqlite>=0.17.0
python-dotenv>=0.19.0
flask>=2.0.0
